Create Longest_Path_Finder.py
Given an M x N matrix, with a few hurdles arbitrarily placed, calculate the length of longest possible route possible from source to destination within the matrix. We are allowed to move to only adjacent cells which are not hurdles. The route cannot contain any diagonal moves and a location once visited in a particular path cannot be visited again.In the given matrix longest path from (0,0)to(1,7) is 24
1 1 1 1 1 1 1 1 1 1
1 1 0 1 1 0 1 1 0 1
1 1 1 1 1 1 1 1 1 1
0-indicates obstacles
1-indicates path
Tailor the matrix for your own custom made maze by changing 1's and 0's